chore(574): Added Elements Dropdown menu
